i am trying to fix some errors calculation of rays casting

+++ include/cub3D.h
@@ -6,7 +6,7 @@
 /*   By: yismaili < yismaili@student.1337.ma>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/21 13:56:15 by yismaili          #+#    #+#             */
-/*   Updated: 2022/11/16 13:16:14 by yismaili         ###   ########.fr       */
+/*   Updated: 2022/11/16 18:14:50 by yismaili         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,8 +53,8 @@ typedef struct  s_player{
 
 typedef struct  s_ray{
 	double rayAngle;
-	int 	wallHit_x;
-	int 	wallHit_y;
+	double 	wallHit_x;
+	double	wallHit_y;
 	double 	Distance;
 	int		hitVertucal;
 	double  rayFacingDown;
